competitor-digest/requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
anthropic>=0.39.0
python-dotenv>=1.0.0
pyyaml>=6.0
playwright>=1.40.0
